# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Riad Lumière Luxury Hotel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List all rooms and suites
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: maxGuests
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get room details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found

  # --- AVAILABILITY ---
...
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      summary: Get restaurant details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}/reserve:
    post:
      operationId: reserveRestaurant
      tags: [restaurants]
      summary: Make a dining reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiningReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiningReservation"

  # --- GALLERY ---
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List guest reviews
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a guest review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # --- NEWSLETTER ---
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterInput"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscription"

  # --- CONTACT ---
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact / concierge inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Inquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"

  # --- ADMIN / DASHBOARD ---
...
components:
  schemas:
...
    Room:
      type: object
      required: [id, name, category, description, pricePerNight, maxGuests, size, images, amenities, available]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          description: "riad-suite | grand-suite | deluxe-room | prestige-room | royal-pavilion"
        description:
          type: string
        shortDescription:
          type: string
        pricePerNight:
          type: number
        maxGuests:
          type: integer
        size:
          type: integer
          description: Square meters
        bedType:
          type: string
        floor:
          type: integer
        images:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        available:
          type: boolean
        featured:
          type: boolean
...
    Restaurant:
      type: object
      required: [id, name, cuisine, description, images]
      properties:
        id:
          type: integer
        name:
          type: string
        cuisine:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        dressCode:
          type: string
        openHours:
          type: string
        capacity:
          type: integer
        chefName:
          type: string
        chefBio:
          type: string
        images:
          type: array
          items:
            type: string
        menuHighlights:
          type: array
          items:
            type: string
        featured:
          type: boolean
    DiningReservationInput:
      type: object
      required: [date, time, guests, guestName, guestEmail]
      properties:
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        specialRequests:
          type: string
    DiningReservation:
      type: object
      required: [id, restaurantId, date, time, guests, guestName, guestEmail, confirmationCode]
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        restaurantName:
          type: string
        date:
          type: string
        time:
          type: string
        guests:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        specialRequests:
          type: string
        confirmationCode:
          type: string
...
    Review:
      type: object
      required: [id, guestName, country, rating, title, body, stayDate, featured]
      properties:
        id:
          type: integer
        guestName:
          type: string
        country:
          type: string
        rating:
          type: number
        title:
          type: string
        body:
          type: string
        stayDate:
          type: string
        roomName:
          type: string
        featured:
          type: boolean
        avatarUrl:
          type: string
    ReviewInput:
      type: object
      required: [guestName, country, rating, title, body, stayDate]
      properties:
        guestName:
          type: string
        country:
          type: string
        rating:
          type: number
        title:
          type: string
        body:
          type: string
        stayDate:
          type: string
        roomName:
          type: string
    NewsletterInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        firstName:
          type: string
    NewsletterSubscription:
      type: object
      required: [id, email, subscribedAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        subscribedAt:
          type: string
    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
        preferredContact:
          type: string
    ContactSubmission:
      type: object
      required: [id, name, email, subject, message, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        subject:
          type: string
        message:
          type: string
        createdAt:
          type: string
...