# lib/api-zod/src/generated/types/newsletterSubscription.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Riad Lumière Luxury Hotel API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscription {
  id: number;
  email: string;
  firstName?: string;
  subscribedAt: string;
}