# lib/db/src/schema/reviews.ts
import { pgTable, serial, text, numeric, boolean, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const reviewsTable = pgTable("reviews", {
  id: serial("id").primaryKey(),
  guestName: text("guest_name").notNull(),
  country: text("country").notNull(),
  rating: numeric("rating", { precision: 3, scale: 1 }).notNull(),
  title: text("title").notNull(),
  body: text("body").notNull(),
  stayDate: text("stay_date").notNull(),
  roomName: text("room_name"),
  featured: boolean("featured").notNull().default(false),
  avatarUrl: text("avatar_url"),
  createdAt: timestamp("created_at").defaultNow(),
});

export const insertReviewSchema = createInsertSchema(reviewsTable).omit({ id: true, createdAt: true });
export type InsertReview = z.infer<typeof insertReviewSchema>;
export type Review = typeof reviewsTable.$inferSelect;